Django/5_school-backend-drf/requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
commonmark==0.9.1
defusedxml==0.7.1
Django==4.1.1
djangorestframework==3.14.0
drf-extensions==0.7.1
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
iniconfig==1.1.1
install==1.3.5
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.12
whitenoise==6.2.0
